+++ RobotNoASF/RobotNoASF.cproj
@@ -235,6 +235,9 @@
     <Compile Include="twi0_interface.h">
       <SubType>compile</SubType>
     </Compile>
+    <Compile Include="twimux_interface.c">
+      <SubType>compile</SubType>
+    </Compile>
     <Compile Include="twimux_interface.h">
       <SubType>compile</SubType>
     </Compile>
